+++ cli/srecli/service/constants.py
@@ -84,6 +84,8 @@ class Endpoint(str, Enum):
     RULE_META_UPDATER = '/rules/update-meta'
     REPORTS_PUSH_DOJO = '/reports/push/dojo'
     CUSTOMERS_RABBITMQ = '/customers/rabbitmq'
+    INTEGRATIONS_EVENT_SOURCES = '/integrations/event-sources'
+    INTEGRATIONS_EVENT_SOURCES_ID = '/integrations/event-sources/{id}'
     REPORTS_DIAGNOSTIC = '/reports/diagnostic'
     REPORTS_DEPARTMENT = '/reports/department'
     INTEGRATIONS_SELF = '/integrations/temp/sre'