+++ cli/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.13.0] - 2026-03-18
+
+### Added
+- Added `sre integrations event_sources` commands: `add`, `describe`, `update`, `delete` for SQS event source management
+- Added default value for the tenant parameter of the command policy add
+
 ## [5.12.0] - 2026-03-02
 
 ### Added